fix(dashboard): satisfy lint on assistants onboarding
Drop unused no-console disable in insights-sidebar and split
useAssistantDraft hook out of AssistantDraftContext so the file
only exports components (react-refresh/only-export-components).
