+++ client/dashboard/src/components/insights-sidebar.tsx
@@ -527,7 +527,6 @@ function PendingPromptBridge({
     try {
       assistantRuntime.thread.append(text);
     } catch (err) {
-      // eslint-disable-next-line no-console
       console.error("Failed to send Explore prompt:", err);
     }
